move gas stats calculation to constructor
